HSEARCH-4089 Use cascading inheritance for ElasticsearchTestDialect implementations
So that we don't duplicate the code so much.
