+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/dialect/Elasticsearch5TestDialect.java
@@ -6,51 +6,13 @@
  */
 package org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-
-import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
-import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-
 import com.google.gson.JsonObject;
 
-@SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
-
-	@Override
-	public boolean isEmptyMappingPossible() {
-		return true;
-	}
-
-	@Override
-	public URLEncodedString getTypeKeywordForNonMappingApi() {
-		return Paths.DOC;
-	}
-
-	@Override
-	public Optional<URLEncodedString> getTypeNameForMappingAndBulkApi() {
-		return Optional.of( Paths.DOC );
-	}
-
-	@Override
-	public Boolean getIncludeTypeNameParameterForMappingApi() {
-		return null;
-	}
-
-	@Override
-	public List<String> getAllLocalDateDefaultMappingFormats() {
-		return Arrays.asList( "yyyy-MM-dd", "yyyyyyyyy-MM-dd" );
-	}
+public class Elasticsearch5TestDialect extends Elasticsearch60TestDialect {
 
 	@Override
 	public void setTemplatePattern(JsonObject object, String pattern) {
 		object.addProperty( "template", pattern );
-	}
-
-	@Override
-	public boolean supportsGeoPointIndexNullAs() {
-		return false;
 	}
 
 	@Override
@@ -59,33 +21,7 @@ public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
 	}
 
 	@Override
-	public boolean zonedDateTimeDocValueHasUTCZoneId() {
-		return true;
-	}
-
-	@Override
-	public boolean supportsIsWriteIndex() {
-		return false;
-	}
-
-	@Override
 	public boolean hasBugForSortMaxOnNegativeFloats() {
 		return true;
-	}
-
-	@Override
-	public boolean hasBugForBigIntegerValuesForDynamicField() {
-		return true;
-	}
-
-	@Override
-	public boolean normalizesStringArgumentToWildcardPredicateForAnalyzedStringField() {
-		return false;
-	}
-
-
-	@Override
-	public boolean supportsSkipOrLimitingTotalHitCount() {
-		return false;
 	}
 }
