+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/dialect/Elasticsearch67TestDialect.java
@@ -13,11 +13,8 @@ import java.util.Optional;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-
 @SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch67TestDialect implements ElasticsearchTestDialect {
+public class Elasticsearch67TestDialect extends Elasticsearch70TestDialect {
 
 	@Override
 	public boolean isEmptyMappingPossible() {
@@ -45,40 +42,8 @@ public class Elasticsearch67TestDialect implements ElasticsearchTestDialect {
 	}
 
 	@Override
-	public void setTemplatePattern(JsonObject object, String pattern) {
-		JsonArray array = new JsonArray();
-		array.add( pattern );
-		object.add( "index_patterns", array );
-	}
-
-	@Override
-	public boolean supportsGeoPointIndexNullAs() {
-		return true;
-	}
-
-	@Override
-	public boolean supportsStrictGreaterThanRangedQueriesOnScaledFloatField() {
-		return true;
-	}
-
-	@Override
 	public boolean zonedDateTimeDocValueHasUTCZoneId() {
 		return true;
-	}
-
-	@Override
-	public boolean supportsIsWriteIndex() {
-		return true;
-	}
-
-	@Override
-	public boolean hasBugForBigIntegerValuesForDynamicField() {
-		return true;
-	}
-
-	@Override
-	public boolean normalizesStringArgumentToWildcardPredicateForAnalyzedStringField() {
-		return false;
 	}
 
 	@Override
