+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/dialect/ElasticsearchTestDialect.java
@@ -64,19 +64,11 @@ public interface ElasticsearchTestDialect {
 
 	boolean supportsIsWriteIndex();
 
-	default boolean hasBugForSortMaxOnNegativeFloats() {
-		return false;
-	}
+	boolean hasBugForSortMaxOnNegativeFloats();
 
-	default boolean hasBugForBigIntegerValuesForDynamicField() {
-		return false;
-	}
+	boolean hasBugForBigIntegerValuesForDynamicField();
 
-	default boolean normalizesStringArgumentToWildcardPredicateForAnalyzedStringField() {
-		return true;
-	}
+	boolean normalizesStringArgumentToWildcardPredicateForAnalyzedStringField();
 
-	default boolean supportsSkipOrLimitingTotalHitCount() {
-		return true;
-	}
+	boolean supportsSkipOrLimitingTotalHitCount();
 }
